[BUG] CLASP logic: remove indexes from exclusion zone that are out of range (#5459)
Fixes #5451.

Sometimes, the `exclusion_zone` array can contain indexes that are out
of the range of `y_pred`. This causes a segmentation fault from numba
when trying to access the elements of `y_pred` using `exclusion_zone`.

To fix this, any indexes that are outside the range of `y_pred` are
removed from `exclusion_zone`.