[BUG] Fix pandas `FutureWarning` for silent upcasting (#5395)
Did not create an issue as this was a very small and trivial fix.

This PR fixes the silent upcasting of the `'window'` column from `int64`
to `object`. This currently (since pandas v2.1.0) raises a deprecation
warning and will raise an error in the future.

See:
https://pandas.pydata.org/docs/whatsnew/v2.1.0.html#deprecated-silent-upcasting-in-setitem-like-series-operations

This is a big issue as it can cause some common IDE's like PyCharm or
VSCode to crash by printing too much output to the terminal at a very
high rate. Even in instances where the IDE does not crash I have
experienced significant slow downs because of this.