[ENH] Add tecator dataset for time series regression as `sktime` onboard dataset (#5428)
#### Reference Issues/PRs
https://github.com/sktime/sktime/issues/4314
#5322 

#### What does this implement/fix? Explain your changes.
It adds a time series regression dataset to the dataset repository,
which is currently missing.

#### Does your contribution introduce a new dependency? If yes, which
one?
No

#### What should a reviewer concentrate their feedback on?
This PR is a direct continuation of #5322, but I don't think any of its
changes made it through to the PR it got merged with.

With that said, here are the changes made from that PR:

- Added citations to the docstring of `load_tecator`
- Added the dataset while avoiding making any changes to the ts
specification. I hadn't realized that @achieveordie had already
accounted for regression datasets with the `targetLabel` tag, which I
wasn't aware of last time.

So overall this is a cleaner implementation since it leaves everything
else alone.

#### Did you add any tests for the change?
I added `load_tecator` to the UNIVARIATE list in
`test_single_loaders.py` so it could be included in the current set of
unit tests for `.ts` files.