[BUG] in `ClaSPSegmentation`, deal with `k` when it is too large for `np.argpartition` (#5490)
Closes #5476.

Sometimes the `ClaSPSegmentation` algorithm raises a value error in the
`_compute_distances_iterative` method. This happens when
`_compute_distances_iterative` attempts to use `np.argpartition` with a
`kth` value which is larger than the length of the input array.

This PR checks if `k` is larger than the length of `dist` and if it is,
it sets `k` to be the length of `dist` minus 1.