[ENH] in `NaiveForecaster`, add valid variance prediction for in-sample forecasts (#5499)
This PR addresses warnings arising from testing `NaiveForecaster`, as
sqrt-s of negative fh-indices are taken.

The reason for this was that the standard deviation formula was taking
sqrt-s of the fh-index, so this is addressed by correcting the formula
for in-sample values.

For this, the error term standard deviation is used with a multiplying
factor of 1. This should be mathematically sensible, and finally silence
the warning.