[BUG] fix `plot_series` prediction interval plotting for 3 or less points in forecasting horizon (#5494)
Fixes https://github.com/sktime/sktime/issues/5493.

The deeper reason was that the `plot_interval` logic seemed broken, it
is unclear why it was working at all even in the case of 4 or more
points. Possibly some fallback functionality in `numpy` which was
kicking in, e.g., using the last indices if broadcast sizes do not
match.

Also adds a test with the failure case from #5493.