[ENH] `iisignature` backend option for `SignatureTransformer` (#5398)
Fixes a bug where a runtime error is returned if each point of the
time-series has dimension more than 25. Fixes #1591 indirectly through the new backend.

Also allows "no augmentation" choice, currently the code returns an error if
augmentations is None -- the user will be forced to choose an
augmentation.
This PR adds enhanced flexibility to allow to user to choose no stream
augmentation.