[ENH] Change GGS to inherit from BaseSeriesAnnotator (#5315)
Updates `GreedyGaussianSegmentation` to inherit from
`BaseSerieAnnotator` following the `HMM` class as a guide.


#### What should a reviewer concentrate their feedback on?

<!-- This section is particularly useful if you have a pull request that
is still in development. You can guide the reviews to focus on the parts
that are ready for their comments. We suggest using bullets (indicated
by * or -) and filled checkboxes [x] here -->

* Currently type check and conversion is done in the `_predict` method.
It would be good to refactor this out into the `BaseSeriesAnnotator`
class but I think that is a job for another PR.
* This PR makes no attempt to convert the `GGS` class to inherit from
`BaseSeriesAnnotator` class. It would be good for this to be done but
again I think this is a job for another PR.

I had to add the type conversions in the `_predict` method to get
the `annotation` tests to pass locally.