fix: pass correct level argument from StatsForecastBackAdapter to statsforecast
Prior to this commit, the 'level' argument being passed to
statsforecast was being subtracted from 1, resulting in extremely narrow
prediction intervals in the common case where the requested `coverage`
was close to 1. For example, calling

```python
model = StatsForecastMSTL(season_length=[12])
model.predict_interval(fh, coverage=0.95)
```

would result in the underlying statsforecast trend model being called
with `level=[5]`, corresponding to a 5% prediction intervl.

This commit removes the subtraction from 1, so that the correct level is
passed.

Fixes #5586.