Revert "[ENN] Improve memory efficiency of `ConformalIntervals` using sparse matrices" (#5609)
Reverts sktime/sktime#5570

I found a bug introduced by this PR. Scipy sparse matrices' default
value is zero and not none. In Line 209, we are fetching the residuals
of our residuals matrix and consider only values that are not Nan
`resids = resids[~np.isnan(resids)]`. Due to the PR, the default values
are not filtered anymore, and we are considering the zeros as normal
residuals. This leads to too sharp intervals.

I think the problem occurs if `sample_freq` is set in a way that not for
all time steps a forecast is made.

I see three possible fixes:
* Revert the PR
* Consider the zero as default value and filter for zeros. However, zero
may also be a valid value. Thus, I think this would introduce an
additional bug.
* We store for which values residuals are calculated. Thus, we need not
to rely on the nan value in line 209.


Thus, I propose to revert the original enhancement.