[ENH] Correct inference of `TransformerPipeline` output type tag (#5625)
Fixes https://github.com/sktime/sktime/issues/5624.

The output tag would not be correctly inferred if a series-to-panel
transformer is followed by series-to-series.
This has been corrected.

Unfortunately, the solution is somewhat involved. I think this hints
that we may want to incode the input/output type information
differently, using number of dimensions or similar.