[ENH] reactivate and fix `test_multiprocessing_idempotent` (#5573)
This PR fixes two untracked issues:

* `test_multiprocessing_idempotent` would be skipped. A comment said
that some issues were related to it, but there was no GitHub issue. This
PR removes the skip, checking whether the problem is gone.
* `test_multiprocessing_idempotent` would also fail for forecasters with
probabilistic capability, due to the output type not being comparable
via the approx array comparison. The modified test now skips for
`predict_proba` outputs.

The failures would also be thrown when using `check_estimator`, this was
noticed by @benheid for `ConformalIntervals`.