[ENH] Support for panel-to-series transformers, merger transformation (#5351)
This PR implements a merger to create a single time series out of a panel. To
enable it , it introduces Transformers that require Panels as input and
transform them to Series. Thus the following adaption to base classes
are needed:
* Determine the output scitype. Thus, in base.py `convert_output` an
elif/else are added.
* In `_convert_output` of base.py, a check is added if the input
conversion was executed based on incompatibilities to the scitype
required by the algorithm or not. If due to incompatibilities, we assume
that the output type of the algorithm should not be changed! Otherwise,
we assume that this conversion is done because of the input data
provided!

Moreover, the expected scitype for the tests is adapted to reflect the
new kind of transformer.