[MNT] refactoring new CI to fix some bugs and other minor enhancements (#5638)
Fixes #5559 (indirectly depends on #5670 for all tests to pass as part
of #5639)

This fixes few bugs (typos, possibly unreliable coverage, etc.) and
added some minor enhancements in new CI.

1. fix main typo: sktime-component was using
detect_components_with_extras which is undefined
2. split test_components into two separate workflows:
1. one to test individual components using dedicated extras and reuse
defined actions
2. one to test all other components without extras using
all_extras_pandas2
3. split long commands into multiple lines using YAML multiline strings
(ref. https://yaml-multiline.info/)
4. skip duplicated testing of base as part of testing components without
dedicated extras
5. remove coverage in partial tests
6. add coverage in full tests if missing
7. added validation of all_extras_pandas2
8. minor changes (mostly renaming) in some places

The changes are tested in #5639 and as it can be seen in
https://github.com/sktime/sktime/actions/runs/7339442367, both
components with and without extras trigger sub-jobs as expected.