[ENH] in `VectorizedDF`, partially decouple internal data store from methods (#5681)
This PR aims to decouple internal methods of `VectorizedDF` as much as
possible from the internal data reference `self.X` and
`self.X_multiindex`. Related to discussion in
https://github.com/sktime/sktime/pull/5676

This is unfortunately not entirely possible, as the central `items`
method needs access to data.

An open question is, how to reconcile this with a requirement that no
data is present when pickling after `fit`.