[ENH] support for probabilistic regressors (`skpro`) in `make_reduction`, direct reduction (#5536)
The aim of this PR is to extend `make_reduction` classes to support
regressors with probabilistic prediction modes, with the `skpro`
interface.

This would be achieved exactly as currently in `YfromX` which already
does support probabilistic regressors, but unfortunately the inner
structure of `make_reduction` is much more nested and organically grown.

The idea is to switch over `predict` and `predict_interval` internally,
and leave the rest as-is.

Hopefully, this will also not interact with the fixes to the reducers
that @benHeid is currently working on.

Depends on #5539, #5725, #5726, which should be merged first.

Also fixes an unreported bug where the reducers could fail in a case
where not all index levels were named, as said names were used as
reference (in `_cut_df`). This has been remedied by using integer
references.