[DOC] updated algorithm inclusion guide (#5753)
This PR updates the algorithm inclusion guide, changes are primarily of
technical nature, due to last year's developments in dependency
management on the level of estimators.

The decision making part is not changed.

Changes relate to:

* explaining that dependencies are managed on estimator level
* therefore additional options are available to power users, such as
maintaining an estimator in a third party codebase. Dependency
management within `sktime` is also more lenient due to depedency
isolation.
* in consequence, first party companion packages like `sktime-dl` are no
longer a primary means to manage soft dependencies, in first party code
this is also addressed in the text. For third party code, no companion
package is necessary either.