[ENH] fix `test_evaluate_error_score` to skip test of expected warning raised if the `joblib` backend is `"loky"` or `"multiprocessing"` (#5780)
Fix `test_evaluate_error_score` to skip test of expected warning raised
if the `joblib` backend is `"loky"` or `"multiprocessing"`. In these
cases, no warnings are raised, this is a known issue:
https://github.com/sktime/sktime/issues/5307

However, the change in the type of the `backend` test fixture has led to
the skip having become ineffective. This PR restores the skip, and adds
a skip for `"multiprocessing"`, which previously was not tested, but has
the same problem.