[ENH] parallelization support and config for forecasting performance metrics (#5813)
This PR adds the option to parallelize forecasting performance metrics
via `set_config`.

The method is simply passing on the config to `vectorize_est`, if set.

Tests are added. Docstrings do not need to be added, as apprently the
option is already inherited from `BaseObject` (mistakenly before this
PR, but now correctly).