[DOC] improved deprecation recipes (#5890)
This PR splits the current deprecation recipe example into two, and adds
more details.

FYI @yarnabrina, I think your example was quite nice, but it combined
two of the "base recipes" (changing default value, changing name), so is
probably going to be confusing to a new developer. It is better if any
examples match one of the recipes 1:1. I think it also was not entirely
in line with the recipe (e.g., use of `None` vs string `"deprecated"`).

Ultimately, I split it into two 1:1 examples for the two base recipes
that the current example was composed of.

Also fixes some indentation issues which made the recipes format
strangely.