[ENH] `ForecastX` option to use future-known variables as exogenous variables in forecasting future-unknown exogenous variables (#5926)
This PR adds to `ForecastX` an option to use future-known variables as
exogeneous variables in forecasting future-unknown exogeneous variables
- implements https://github.com/sktime/sktime/issues/5917.

Related: @yarnabrina's use case from
https://github.com/sktime/sktime/discussions/5903 where this should be
helpful.

Minor additional changes:

* docstring formatting
* adds test coverage for parameter `fit_behaviour` - this was apparently
missing? FYI @benHeid
* allows `forecaster_X` to be optional, in this case `forecater_y` will
be used

Also fixes an unreported bug with the `fit_behaviour="use_forecast"`
option. If internally, `X.index` as integer, it would be interpreted as
relative forecast ahead istead of - as intended - an in-sample forecast
for the indices of `X`, which would lead to dependent failures.