[ENH] Add nan policy handler for FallbackForecaster #5897 (#5924)
<!--
Welcome to sktime, and thanks for contributing!
Please have a look at our contribution guide:
https://www.sktime.net/en/latest/get_involved/contributing.html
-->

#### Reference Issues/PRs
-Issue: https://github.com/sktime/sktime/issues/5897
-Additional discussion: https://github.com/sktime/sktime/pull/5847


#### What does this implement/fix? Explain your changes.

We've introduced a `nan_predict_policy` within the `FallbackForecaster`
class to handle NaN values in forecaster predictions. This policy
defaults to "ignore" and can be set to "raise", which will actively
check for NaN values in the prediction output and raise an exception if
any are detected, or "warn", which will warn the user if any values are
detected. This enhancement is aimed at improving the robustness and
reliability of the forecasting process, especially when integrating
transformations like BoxCox with ARIMA forecasters.