[ENH] Expose seasonality parameters of ProphetPiecewiseLinearTrendForecaster (#5834)
<!--
Welcome to sktime, and thanks for contributing!
Please have a look at our contribution guide:
https://www.sktime.net/en/latest/get_involved/contributing.html
-->
The PR exposes the seasonal parameters of the
ProphetPiecewiseLinearTrendForecaster.

The PR is the result of the discussion in this
(https://github.com/sktime/sktime/pull/5592).

#### What does this implement/fix? Explain your changes.
<!--
A clear and concise description of what you have implemented.
-->
This is a simple change to allow the user to define what the seasonality
parameters should be. @tpvasconcelos suggested they should be

```python
daily_seasonality=False,
weekly_seasonality=False,
yearly_seasonality=False
```
but as the discussion (https://github.com/sktime/sktime/pull/5592)
showed it is not clear if this is the best setting.