[ENH] Cinn based forecaster (#5339)
## Note relies on 
* DeepLearning base design is merged 
* CurveFitForecaster is merged
* Merger #5351 

#### Reference Issues/PRs
Fixes #4924

#### What does this implement/fix? Explain your changes.
It implements the method of the BigDEAL Challenge that won the peak
shape and the peak timing forecasting tasks.

Furthermore, it adapts a bit the base classes for torch forecasters to
make them more flexible.

#### Does your contribution introduce a new dependency? If yes, which
one?
It introduces FrEIA as a dependency. This is a framework for
implementing reversible neural network architectures.