[ENH] Add neuralforecast adapter and rnn forecaster (#5962)
Adapter class to `neuralforecast` from `Nixtla` and concrete estimator
for `RNN`.

Notes for reviewers (not in any particular order)

1. `neuralforecast` supports 3 types of exogenous features: static
(constant over time), historical (unknown for prediction horizon) and
future (known for prediction horizon). I felt the static type has no
predictive capability if models are trained per series, and as far as I
know, sktime will not be able to support historical type for framework
compatibilities. Hence in this first draft of adaptation, I have added
support for only future type.

2. `neuralforecast` has the capability of directly supporting
panel/hierarchical datasets directly. I am not sure how to take benefit
of that in sktime, so have not used it in this first draft. If it's easy
enough to support, I can update if some references are shared, otherwise
this can definitely be a extension.

3. ~`RNN` in `neuralforecast` supports specification of loss and
validation loss, but they needs to be instances of [pytorch loss
functions](https://nixtlaverse.nixtla.io/neuralforecast/losses.pytorch.html).
I am not sure how set these defaults in instance, as soft dependencies
can not be checked there. I have hence not exposed these parameters and
rely on the defaults, but this is both important and common parameter.
With some suggestions and references, I will like to enhance this in
this draft itself.~ Added some support using internal attributes.

4. `neuralforecast` does support probablistic forecasts, but it seems
that depends on choice of loss function (and may be something more, not
sure). In this first draft, I have not added that support so far and
this can be part of an extension.

5. `neuralforecast` has a public `NeuralForecast` class which accepts
`pandas.DataFrame` objects of special schema directly, and has the
ability to train multiple models and find best one. The underlying
models do not support `pandas.DataFrame` directly and relies of
`TimeSeriesDataset`, a sub-class of `torch.utils.data.Dataset`. For
simplification with most other forecasters, I opted for the higher level
class.

6. `neuralforecast` supports in-sample predictions, but warns that those
may be inaccurate. Since the library itself is marked as `Alpha`, I have
disabled entire support of in-sample predictions for now.

7. `neuralforecast` takes a lot of dependencies, and adding it in
`forecasting` extra may cause too much installation for users who won't
use it. But adding it in `dl` extra means new CI is not going to test
it, so we need to decide what to do.

8. ~`neuralforecast` supports a lot of models, and (hopefully) all of
them can be used with the adapter in this PR. However that is a lot of
models, and it leads to the question of how should we organise them, by
model or by dependency or by both or by something else.~ Modified only
`dl` dependency similar to other recent PR's that affect
torch/tensorflow etc.