Add a `test_mstl` module checking if `transform` returns desired components (#6084)
Closes #6083, resolves #5458.

This PR adds a test file for the
[MSTL](https://github.com/sktime/sktime/blob/main/sktime/transformations/series/detrend/mstl.py)
class. Currently, all it intends to do is test whether
[MSTL](https://github.com/sktime/sktime/blob/main/sktime/transformations/series/detrend/mstl.py)'s
`transform` returns the correct components when `return_components` is
set to True during initialization.

If it would be beneficial, I could probably expand the testing script to
test MSTL in other scenarios (like we do for `_detrender` and
`_deseasonalizer`). From @fkiraly's last comment in #5458, I thought
this would be the most immediately helpful and nice-to-have test.

There's just one method in my code right now so it should be quick to
review. This is my first PR here and I'm still familiarizing myself with
the codebase, so any feedback would help.