[BUG] Ignore `ValueError` on `pd.infer_freq` when index has fewer than 3 values (#6097)
#### Reference Issues/PRs

Solves #5853

#### What does this implement/fix? Explain your changes.

Avoid propagating a ValueError to the end-user when the `pd.infer_freq`
can't infer the frequency of the passed series. This happens when the
series has fewer than 3 elements. This patch catches the exception and
falls back to `None` for the series' frequency value.

See examples discussed in #5853 and in the tests.