[BUG] search function for estimator overview not working (#6105)
Fixes #5495

I identified the location of the Js search function. It was written with
jQuery which I was not familiar with. I just rewrote it in vanilla Js
and the problem was fixed.
 
This is very basic so it may not but the most right or efficient. I am
going to look further into the documentation and implement some other
enhancements for the estimator overview in particular and the docs in
general. Then I may find out why the jQuery didn't work or draw up some
other solutions.