[MNT] revert switch from `pykalman` to `pykalman-bardo` (#6114)
Update: @mbalatsko has now merged the 3.11/3.12 compatibility fixes back
into `pykalman`.

With that a simpler solution is possible: simple revert, allow all
python versions, and no `backend` parameter is needed.


---
About half a year ago, this PR changed the dependency `pykalman` to
`pykalman-bardo`: https://github.com/sktime/sktime/pull/5277.

We had a tracking issue to resolve the situation:
https://github.com/sktime/sktime/issues/5414

Since then:

* author @mbalatsko became an owner of `pykalman`, see
https://github.com/pypi/support/issues/3196
* after that, nothing happened with `pykalman`, or `pykalman-bardo` - in
particular, the python 3.11/3.12 patch has not been played back to
`pykalman`, nor was there any other interaction of @mbalatsko with the
`pykalman` repository
* request for status update has remained unanswered for 3 months:
https://github.com/sktime/sktime/issues/5414#issuecomment-1835728124
* due to the unclear maintenance situation, multiple users have
published their own forks of `pykalman` on `pypi` with various fixes,
but no consensus fork has emerged.

I would hence propose a simple revert of #5277.

I have opened an issue on `pykalman` to clarify the situation - my
baseline expectation is that there is no reply by maintainers:
https://github.com/pykalman/pykalman/issues/109

What is also worth consideration to which extent this situation
constitutes a security risk for users. The extras are not that
frequently downloaded, but even a small fraction of total `sktime`
downloads is still a substantial number.

Beyond the revert, I have also added a `backend` parameter to the
estimator, for checking purposes, and not restricted python versions.
This allows users to choose their own backend. The error message will
not fire, and the estimator should work, as long as one of the many
forks of `pykalman` is installed.