[DOC] Miscellaneous aesthetic improvements to docs UI (#6211)
#### What does this implement/fix? Explain your changes.

- [ ] Add "theme-switcher" to navbar: Allow to switch between dark and
light mode.
- [ ] Remove search-field from left sidebar: The search field on left
sidebar is confusing in term of UX. When users click on the search bar
on the top right, the cursor instead jumps to the left search bar. The
default global search overlay is better and used by all docs with
pydata-theme, also considering that search directs to a new page.
- [ ] Change some icons with white background to transparent. They are
apparent in dark mode. I happened to have a quick script for that.
- [ ] Remove some empty left sidebars like of homepage. 

If the PR seems inconsequent I am sorry. I made these changes when
testing the docs and still left them around. Figure there is no harm in
some aesthetic improvement. If any changes not approved i would just
revert them.