[ENH] Option to use future-unknown exogenous variables in `ForecastX` if passed in `predict` (#6199)
Before this PR, all predict methods of `ForecastX` always used
`forecaster_X` to predict `X` first and use that as input to
`forecaster_y`. This used to happen regardless of the fact whether those
columns were passed in passed value of `X` argument.

This PR gives an option to control this behaviour. If user passes
`predict_behaviour` as `"only_if_necessary"`, `X` will not be predicted
if all variables specified in `self.columns` are present in the user
input already. By default, this argument uses `"always"` to keep
existing behaviour.