[MNT] fix isolation of `mlflow` soft dependencies (#6285)
It seems the fix #6282 caused another issue - I believe the reason is
that `dask` removed `pyyaml` as a dependency.

Consequently, the `import yaml` statement in the mlflow module now
fails, post #6282. Unclear why the tests did not recognize this.

The fix should be this: import `yaml` only if `mlflow` is present.