[ENH] refactor test scenario creation to be lazy rather than on module load (#6278)
This PR refactors test scenario creation to be lazy rather than on
module load.

Currently, the data for test scenarios is created on module load and is
persisted as module variables, which may increase test collection time
and memory usage. The status quo also creates a risk for side effects to
which `pandas` objects are prone.

This PR migrates the architecture to the same pattern that `pytest`
fixtures use, where concrete objects are only created when the scenario
is loaded. The change also ensures that a new object is created each
time the fixture is requested, avoiding side effects from mutating the
single reference.