[DOC] Basic fix and enhancement of doc local build process (#6128)
#### Reference Issues/PRs

#6108  #6106

#### What does this implement/fix? Explain your changes.

This PR improves the local documentation build process to be more
consistent with the online documentation and reduce build times for
better observability.

1. Move Makefile into the source folder, as some functions of conf.py
should be run inside source folder like in Readthedoc as stated in
.readthedocs.yml.

2. Removed `sphinx-apidoc` and related code as it was not configured in
conf.py and so was not used by the build process on Readthedocs. This
eliminates orphan pages in the local build documentation that are not
present in the online documentation.
For example:

![image](https://github.com/sktime/sktime/assets/56506156/c78e048e-37e3-4cc5-9046-f02c02698ba1)
which does not appear in online doc:
https://www.sktime.net/en/stable/api_reference/auto_generated/sktime.alignment.base.html

3. No longer remove build folder. Instead, add a clean target for
situations where a clean build is necessary. This change allows
sphinx-build to detect changes in both the documentation and the
codebase and do its increment build from cache.

4. Commented out `symlink_examples.sh`. Currently each example in docs
is manually linked to sktime/examples. We must refractor the
sktime/examples folder to the same structure assumed by docs before we
could link it. This change is made in #6210.

5. Added an autobuild target using sphinx-autobuild. 

6. Modify docs page on building documentation locally to reflect the
changes

The new Makefile significantly reduces build times for incremental
changes, about 10 times faster than before.

#### Does your contribution introduce a new dependency? If yes, which
one?

No. But users could independently install `sphinx-autobuild` if needed. 

#### What should a reviewer concentrate their feedback on?

This PR won't have any effect on the online documentation. Please
confirm on the efficient and consistent of local documentation build
with online docs if you have time.

#### Any other comments?

> 4. Commented out `symlink_examples.sh`.

The `symlink_examples.sh` file served no purpose, even before this PR,
as the symlinks to examples are tracked. I keep it in the repo in case
we need to dynamically create symlink to docs in the future.