[ENH] enhancements to tag system, systematic API docs for tags (#6289)
This PR upgrades the tag system with improved inspectability, and adds
systematic API docs for the tags.

The changes are as following:

* tags are now changed to be (private) first-class objects. This allows
to attach more extensive documentation to tags, including examples or
references. It also allows to attach tags to tags, e.g., the meta-tag
fields currently in `registry._tags`, or a property such as whether a
tag is meant to be user facing.
* longer-term, it could also allow to attach validation functionality to
a tag, in the form of methods, or enable dispatch, or nested dispatch.
* an API reference page for tags is added. This (ab)uses sphinx autodoc
to generate tables and individual glossary pages for tags, by using the
class template.