[BUG] fix dependent tags of `TimeSeriesDBSCAN` (#6322)
`TimeSeriesDBSCAN` does not set its tags correctly if one of the `numba`
based distance in `sktime` is used. In this case, it cannot support
unequal length time series, and the corresponding tag needs to be set.

Similarly, the tags for some of the `numba` distances were set
incorrectly, which is now corrected as well.