[DOC] author credits to `pycatch22` authors, fix missing documentation page (#6300)
With increased visibility of the authors field in the docs (e.g., in the
reworked [estimator
overview](https://www.sktime.net/en/latest/estimator_overview.html)), we
should make sure it properly reflects contribution. Also see #5938.

This adds upstream authors to the author field of the `pycatch22` facing
transformation, and gives clearer credit to `pycatch22` in the docstring
by naming it.

It also appears that the estimator was never added to the public
documentation page, and only the copy/reimplementation was visible there
- I apologize for that, this dates back to old history of `sktime`, and
was certainly not intentional on my part.

The author credits add @benfulcher and @jmoo2880. The maintainer tag had
@benfulcher, so based on recent contribution activity I added @jmoo2880.
Please correct me if this assumption of co-maintainership is wrong.