[MNT] PEP 508 environment markers for estimators (#6144)
This PR adds functionality for adding PEP 508 environment markers for
specifying estimator dependencies.

This is useful to specify operating systems with which the estimator is
not compatible, e.g., `platform_system!="Windows"`.
The information is added as a string-valued tag `env_marker`, which must
be PEP 508 compliant.

The logic is tied into `_check_estimator_deps`, so estimators marked
with operating system excludes will not be tested on the respective
operating system, and will raise a user friendly error message pointing
to its requirements.

Minor change: moved default values for all current packaging tags
related to the env up to the `BaseObject`. Later we could move this into
`scikit-base`.