Make deep classifier's convert_y_to_keras private
The method `convert_y_to_keras` refits the classifiers' onehot and
label encoder based on the input data. It is automatically called
during `deep_classifier.fit`. If an unwitting user calls it again with
data that has different classes, the behavior of the predict function
will change.

Making the method private prevents accidental incorrect usage and
simplifies the interface.