[ENH] `ExpandingCutoffSplitter` - splitter by moving cutoff (#6360)
#### Reference Issues/PRs

https://github.com/sktime/sktime/discussions/6327
https://github.com/sktime/sktime/issues/6359

#### What does this implement/fix? Explain your changes.

This PR introduces a new splitter, `ExpandingCutoffSplitter`, to the
time series forecasting module of sktime. This splitter starts with an
initial training window based on a specified cutoff point and expands
incrementally in each split. This feature addresses the need for a
dynamic adjustment of the training window in time series forecasting
based on a cutoff date, especially in cases where many different time
series will be tested on common cutoff dates but otherwise have
different starting points.