[BUG] Modified VAR code to allow predict_quantiles of 0.5 (fixes #4742) (#6441)
#### Reference Issues/PRs
Fixes #4742


#### What does this implement/fix? Explain your changes.
Previously, VAR produced an error when predicting quantiles with an
alpha of 0.5. This issue arose because alpha was converted to a coverage
of 1.0 in `_base`, which was then reconverted to an alpha ("internal
alpha") of 1.0 inside VAR's `_predict_interval`. An internal alpha of
1.0 is outside the acceptable range (0.0 to 1.0 exclusive) for
statsmodels VAR.

This change hackily coerces an internal alpha of 1.0 to 0.99999 to avoid
this error.