Deprecate BaseDeepClassifier convert_y_to_keras
`convert_y_to_keras` should not be a public method, as disDeprecate
BaseDeepClassifier convert_y_to_keras

`convert_y_to_keras` should not be a public method, as explained in
d1a802985e3a4e14c4f06ebedd0c083fd63c71a6. This change issues a
deprecation warning when the method is called and delegates to the
private method, which is used by all deriving classes. The deprecation
message starts with "WARNING" to match the other deprecation message in
the modified file.