[MNT] run tests in `distances` module only if it has changed (#6517)
This PR adds a switch that runs tests in the `distances` module only if
it has changed. This will have a substantial impact on the test times,
as this module adds up to lengthy testing.

Depends on https://github.com/sktime/sktime/pull/6518.

This PR *will* run tests in the `distances` module, as it has changed by
virtue of introducing the pytest conditional testing decorators.