&fkiraly [ENH] enhanced estimator overview table - tag display and search (#6147)
#### Reference Issues/PRs

#6105 #6106 #4905

The additional build on RTD can be found here in case it is not yet
rendered on sktime own RTD.
https://duydlsktime.readthedocs.io/en/latest/estimator_overview.html

#### What does this implement/fix? Explain your changes.
<!--
A clear and concise description of what you have implemented.
-->

An enhancement to estimator overview table. Support filtering by
estimator type and displaying tag.

#### Does your contribution introduce a new dependency? If yes, which
one?

It is implemented in vanilla js.