[ENH] indexing `prophetverse` forecaster (#6614)
This PR adds indexing for the `Prophetverse` forecaster (from `prophetverse` directly in `sktime`, enabling:

* `all_estimators` discovery
* display in the search tables

This is done by adding a placeholder with:

* docstring
* tags
* init with parameter names

I.e., all the database information required and queried for indexing. If `prophetverse` is present, the imports is replaced by a dropin of the forecaster from `prophetverse` (or, by a delegator of reduced functionality if `prophetverse<=0.2.0` is present).