[BUG] minirocket: zero division errors #5174 (#6612)
#### Reference Issues/PRs

Fixes #5174.


#### What does this implement/fix? Explain your changes.

MultiRocket and MiniRocket transforms are not well defined when the
parameter num_kernels is less than 84.

The reason at least 84 kernels must be chosen is as follows. Multi/mini
rocket transforms rely on precisely 84 "base" kernels. Kernels used to
transform the data are generated as dilations of the base 84. When a
user specifies the parameter num_kernels, the algorithm generates a
number of kernels equal to the nearest multiple of 84 (rounding down),
since the number of dilations per kernel is num_kernels // 84. If the
parameter num_kernels is less than 84, no transformation kernels can be
generated, since the number of dilations per kernel is 0.

The authors did not foresee this issue, since they always set the
parameter num_kernels to be in the thousands and recommended this
practice to subsequent users.

When the num_kernels is less than 84 for multi/mini rocket transforms,
the actual number of kernels used is set to 84. A new attribute called
num_kernels_ gives the actual number of kernels to be used. For
transforms, it is set during _fit. For classifiers and regressors using
the "minirocket" or "multirocket" transforms, it is calculated at
initialization.


#### Did you add any tests for the change?

In _rocket_classifier.py and _rocket_regressor.py, some failing
parameter values were added to get_test_params: when the parameter
rocket_transform is either "minirocket" or "multirocket" and num_kernels
< 84, there should no longer be a division by zero error.