[DOC] rst roadmap documentation page stale since 2021 - replace by correct links to recent roadmaps (#6556)
I noticed that the roadmap documentation page still showed the 2021
roadmap - this is because since then planning moved to hackmd/wikis and
then to issues, and apparently we forgot about that page.

To make management of the roadmap easier in the future and avoid
forgetting about it in the docs, I've replaced the 2021 roadmap page
with a link to issues tagged "roadmap". This will direct visitors of the
docs to the place where the roadmap is actually being developed.