[ENH] fitted parameter forwarding utility, forward `statsforecast` estimators' fitted parameters (#6349)
This PR adds a fitted parameter forwarding utility which is useful for
adapter classes. It copies fitted parameters in the sense of the
`sklearn` convention that these end in underscore from any object -
usually the delegate, nthe adapter case - to a target object - usually
`self`.

Further, the utility is added to adapters as follows:

* adds fitted parameter forwarding in
`_GeneralisedStatsforecastAdapter`. This fixes
https://github.com/sktime/sktime/issues/5720
* replaces the ad-hoc code in `_PytsAdapter` with the utility
* replaces the ad-hoc code in `_TslearnAdapter` with the utility